[RENAME] CustomResponse위치변경
